# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sayenly API Documentation",
    "version": "1.0.0",
    "description": "Documentation for Sayenly Backend API",
    "contact": {
      "name": "Sayenly Support",
      "email": "support@sayenly.com"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "FAQ": {
        "type": "object",
        "required": [
          "id",
          "question",
          "answer",
          "inApp",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "The auto-generated ID of the FAQ"
          },
          "question": {
            "type": "string",
            "description": "The frequently asked question"
          },
          "answer": {
            "type": "string",
            "description": "The answer to the question"
          },
          "inApp": {
            "type": "boolean",
            "description": "Whether the FAQ is visible in the app"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "The date the FAQ was created"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "The date the FAQ was last updated"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/faq/get-all": {
      "get": {
        "summary": "Get all FAQs",
        "tags": [
          "FAQ"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of all FAQs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FAQ"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/faq/get-by-id": {
      "get": {
        "summary": "Get a FAQ by ID",
        "tags": [
          "FAQ"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The FAQ ID"
          }
        ],
        "responses": {
          "200": {
            "description": "The FAQ information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FAQ"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "FAQ not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/faq/create": {
      "post": {
        "summary": "Create a new FAQ",
        "tags": [
          "FAQ"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "question",
                  "answer"
                ],
                "properties": {
                  "question": {
                    "type": "string",
                    "description": "The frequently asked question"
                  },
                  "answer": {
                    "type": "string",
                    "description": "The answer to the question"
                  },
                  "inApp": {
                    "type": "boolean",
                    "default": false,
                    "description": "Whether the FAQ is visible in the app"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The created FAQ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FAQ"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/faq/update": {
      "put": {
        "summary": "Update a FAQ",
        "tags": [
          "FAQ"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The FAQ ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "question": {
                    "type": "string",
                    "description": "The frequently asked question"
                  },
                  "answer": {
                    "type": "string",
                    "description": "The answer to the question"
                  },
                  "inApp": {
                    "type": "boolean",
                    "description": "Whether the FAQ is visible in the app"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated FAQ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FAQ"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "FAQ not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/faq/delete": {
      "delete": {
        "summary": "Delete a FAQ",
        "tags": [
          "FAQ"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The FAQ ID"
          }
        ],
        "responses": {
          "200": {
            "description": "FAQ deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "FAQ not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  }
}